+++ FirstCPPApplication/attribute_container.cpp
@@ -64,25 +64,25 @@ std::vector<std::string> AttributeContainer::PrettyVector()
     std::string result = "";
     std::vector<std::string> string_vec;
 
-    string_vec.push_back("HCV: "+std::to_string((long double)this->health->current_val));
-    string_vec.push_back("HMV: "+std::to_string((long double)this->health->max_val));
-    string_vec.push_back("HRR: "+std::to_string((long double)this->health->regen_rate));
-    string_vec.push_back("HRI: "+std::to_string((long double)this->health->regen_interval));
+    string_vec.push_back("HCV: "+std::to_string((long double)(int)this->health->current_val));
+    string_vec.push_back("HMV: "+std::to_string((long double)(int)this->health->max_val));
+    string_vec.push_back("HRR: "+std::to_string((long double)(int)this->health->regen_rate));
+    string_vec.push_back("HRI: "+std::to_string((long double)(int)this->health->regen_interval));
 
-    string_vec.push_back("MCV: "+std::to_string((long double)this->mana->current_val));
-    string_vec.push_back("MMV: "+std::to_string((long double)this->mana->max_val));
-    string_vec.push_back("MRR: "+std::to_string((long double)this->mana->regen_rate));
-    string_vec.push_back("MRI: "+std::to_string((long double)this->mana->regen_interval));
+    string_vec.push_back("MCV: "+std::to_string((long double)(int)this->mana->current_val));
+    string_vec.push_back("MMV: "+std::to_string((long double)(int)this->mana->max_val));
+    string_vec.push_back("MRR: "+std::to_string((long double)(int)this->mana->regen_rate));
+    string_vec.push_back("MRI: "+std::to_string((long double)(int)this->mana->regen_interval));
 
-    string_vec.push_back("ACV: "+std::to_string((long double)this->armor->current_val));
-    string_vec.push_back("AMV: "+std::to_string((long double)this->armor->max_val));
-    string_vec.push_back("ARR: "+std::to_string((long double)this->armor->regen_rate));
-    string_vec.push_back("ARI: "+std::to_string((long double)this->armor->regen_interval));
+    string_vec.push_back("ACV: "+std::to_string((long double)(int)this->armor->current_val));
+    string_vec.push_back("AMV: "+std::to_string((long double)(int)this->armor->max_val));
+    string_vec.push_back("ARR: "+std::to_string((long double)(int)this->armor->regen_rate));
+    string_vec.push_back("ARI: "+std::to_string((long double)(int)this->armor->regen_interval));
 
-    string_vec.push_back("DCV: "+std::to_string((long double)this->damage->current_val));
-    string_vec.push_back("DMV: "+std::to_string((long double)this->damage->max_val));
-    string_vec.push_back("DRR: "+std::to_string((long double)this->damage->regen_rate));
-    string_vec.push_back("DRI: "+std::to_string((long double)this->damage->regen_interval));
+    string_vec.push_back("DCV: "+std::to_string((long double)(int)this->damage->current_val));
+    string_vec.push_back("DMV: "+std::to_string((long double)(int)this->damage->max_val));
+    string_vec.push_back("DRR: "+std::to_string((long double)(int)this->damage->regen_rate));
+    string_vec.push_back("DRI: "+std::to_string((long double)(int)this->damage->regen_interval));
 
     return string_vec;
 };
