+++ FirstCPPApplication/item_effect.cpp
@@ -481,11 +481,11 @@ std::string ItemEffect::full_str()
 
 };
 
-std::string ItemEffect::small_convert(std::string prefix, long double val)
+std::string ItemEffect::small_convert(std::string prefix, int val)
 {
     if (val != 0)
     {
-        return prefix + std::to_string(val);
+        return prefix + std::to_string((long double)val);
     }
     else
     {
@@ -548,36 +548,31 @@ std::vector<TCOD_colctrl_t> ItemEffect::oneline_str_colours()
     
 };
 
-// std::string ItemEffect::oneline_str()
-// {
-// 
-// };
-
 std::string ItemEffect::oneline_str()
 {
 
     std::string result = "";
     std::vector<std::string> string_vec;
 
-    string_vec.push_back(this->small_convert("%cHCV: ", (long double)this->health_current_val));
-    string_vec.push_back(this->small_convert("%cHMV: ", (long double)this->health_max_val));
-    string_vec.push_back(this->small_convert("%cHRR: ", (long double)this->health_regen_rate));
-    string_vec.push_back(this->small_convert("%cHRI: ", (long double)this->health_regen_interval));
+    string_vec.push_back(this->small_convert("%cHCV: ", (int)this->health_current_val));
+    string_vec.push_back(this->small_convert("%cHMV: ", (int)this->health_max_val));
+    string_vec.push_back(this->small_convert("%cHRR: ", (int)this->health_regen_rate));
+    string_vec.push_back(this->small_convert("%cHRI: ", (int)this->health_regen_interval));
 
-    string_vec.push_back(this->small_convert("%cMCV: ", (long double)this->mana_current_val));
-    string_vec.push_back(this->small_convert("%cMMV: ", (long double)this->mana_max_val));
-    string_vec.push_back(this->small_convert("%cMRR: ", (long double)this->mana_regen_rate));
-    string_vec.push_back(this->small_convert("%cMRI: ", (long double)this->mana_regen_interval));
+    string_vec.push_back(this->small_convert("%cMCV: ", (int)this->mana_current_val));
+    string_vec.push_back(this->small_convert("%cMMV: ", (int)this->mana_max_val));
+    string_vec.push_back(this->small_convert("%cMRR: ", (int)this->mana_regen_rate));
+    string_vec.push_back(this->small_convert("%cMRI: ", (int)this->mana_regen_interval));
 
-    string_vec.push_back(this->small_convert("%cACV: ", (long double)this->armor_current_val));
-    string_vec.push_back(this->small_convert("%cAMV: ", (long double)this->armor_max_val));
-    string_vec.push_back(this->small_convert("%cARR: ", (long double)this->armor_regen_rate));
-    string_vec.push_back(this->small_convert("%cARI: ", (long double)this->armor_regen_interval));
+    string_vec.push_back(this->small_convert("%cACV: ", (int)this->armor_current_val));
+    string_vec.push_back(this->small_convert("%cAMV: ", (int)this->armor_max_val));
+    string_vec.push_back(this->small_convert("%cARR: ", (int)this->armor_regen_rate));
+    string_vec.push_back(this->small_convert("%cARI: ", (int)this->armor_regen_interval));
 
-    string_vec.push_back(this->small_convert("%cDCV: ", (long double)this->damage_current_val));
-    string_vec.push_back(this->small_convert("%cDMV: ", (long double)this->damage_max_val));
-    string_vec.push_back(this->small_convert("%cDRR: ", (long double)this->damage_regen_rate));
-    string_vec.push_back(this->small_convert("%cDRI: ", (long double)this->damage_regen_interval));
+    string_vec.push_back(this->small_convert("%cDCV: ", (int)this->damage_current_val));
+    string_vec.push_back(this->small_convert("%cDMV: ", (int)this->damage_max_val));
+    string_vec.push_back(this->small_convert("%cDRR: ", (int)this->damage_regen_rate));
+    string_vec.push_back(this->small_convert("%cDRI: ", (int)this->damage_regen_interval));
 
     return StringJoin(string_vec, ' ', true);
